requirements.txt
matplotlib==3.7.1
numpy==1.23.5
opencv-python==4.7.0.72
tensorflow==2.12.0
fastapi==0.95.1
uvicorn==0.21.1
Pillow==9.5.0
imageio==2.27.0